Refactor agent API skeleton structure
