+++ src/WebApp/PingMonitor.Web/Contracts/Config/AgentConfigResponse.cs
@@ -1,0 +1,6 @@
+namespace PingMonitor.Web.Contracts.Config;
+
+public sealed record AgentConfigResponse(
+    string ConfigVersion,
+    DateTimeOffset GeneratedAtUtc,
+    IReadOnlyList<MonitorAssignmentDto> Assignments);
