+++ src/WebApp/PingMonitor.Web/Contracts/Heartbeat/AgentHeartbeatRequest.cs
@@ -1,0 +1,9 @@
+namespace PingMonitor.Web.Contracts.Heartbeat;
+
+public sealed record AgentHeartbeatRequest(
+    string AgentVersion,
+    DateTimeOffset SentAtUtc,
+    string ConfigVersion,
+    int ActiveAssignments,
+    int QueuedResultCount,
+    string Status);
