+++ src/WebApp/PingMonitor.Web/Contracts/Heartbeat/AgentHeartbeatResponse.cs
@@ -1,0 +1,6 @@
+namespace PingMonitor.Web.Contracts.Heartbeat;
+
+public sealed record AgentHeartbeatResponse(
+    bool Ok,
+    DateTimeOffset ServerTimeUtc,
+    bool ConfigChanged);
