+++ src/WebApp/PingMonitor.Web/Contracts/Hello/AgentHelloRequest.cs
@@ -1,0 +1,8 @@
+namespace PingMonitor.Web.Contracts.Hello;
+
+public sealed record AgentHelloRequest(
+    string AgentVersion,
+    string MachineName,
+    string Platform,
+    IReadOnlyList<string> Capabilities,
+    DateTimeOffset StartedAtUtc);
