+++ src/WebApp/PingMonitor.Web/Models/AgentHealthStatus.cs
@@ -1,0 +1,8 @@
+namespace PingMonitor.Web.Models;
+
+public enum AgentHealthStatus
+{
+    Online,
+    Stale,
+    Offline
+}
