+++ src/WebApp/PingMonitor.Web/Models/AlertStatus.cs
@@ -1,0 +1,7 @@
+namespace PingMonitor.Web.Models;
+
+public enum AlertStatus
+{
+    Open,
+    Closed
+}
