+++ src/WebApp/PingMonitor.Web/Models/CheckType.cs
@@ -1,0 +1,6 @@
+namespace PingMonitor.Web.Models;
+
+public enum CheckType
+{
+    Icmp
+}
