+++ src/WebApp/PingMonitor.Web/Models/EndpointStateKind.cs
@@ -1,0 +1,10 @@
+namespace PingMonitor.Web.Models;
+
+public enum EndpointStateKind
+{
+    Unknown,
+    Up,
+    Degraded,
+    Down,
+    Suppressed
+}
