--- src/WebApp/PingMonitor.Web/Models/Enums.cs
@@ -1,28 +1,0 @@
-namespace PingMonitor.Web.Models;
-
-public enum EndpointStateKind
-{
-    Unknown,
-    Up,
-    Degraded,
-    Down,
-    Suppressed
-}
-
-public enum AlertStatus
-{
-    Open,
-    Closed
-}
-
-public enum AgentHealthStatus
-{
-    Online,
-    Stale,
-    Offline
-}
-
-public enum CheckType
-{
-    Icmp
-}
