+++ src/WebApp/PingMonitor.Web/Services/IAgentAuthenticationService.cs
@@ -1,0 +1,6 @@
+namespace PingMonitor.Web.Services;
+
+public interface IAgentAuthenticationService
+{
+    Task<bool> ValidateAsync(string instanceId, string? authorizationHeader, CancellationToken cancellationToken);
+}
