+++ src/WebApp/PingMonitor.Web/Services/IHeartbeatService.cs
@@ -1,0 +1,8 @@
+using PingMonitor.Web.Contracts.Heartbeat;
+
+namespace PingMonitor.Web.Services;
+
+public interface IHeartbeatService
+{
+    Task<AgentHeartbeatResponse> ProcessHeartbeatAsync(string instanceId, AgentHeartbeatRequest request, CancellationToken cancellationToken);
+}
