+++ src/WebApp/PingMonitor.Web/Services/IStateEvaluationService.cs
@@ -1,0 +1,6 @@
+namespace PingMonitor.Web.Services;
+
+public interface IStateEvaluationService
+{
+    Task EvaluateAssignmentStateAsync(string assignmentId, CancellationToken cancellationToken);
+}
