+++ src/WebApp/PingMonitor.Web/Services/PlaceholderStateEvaluationService.cs
@@ -1,0 +1,9 @@
+namespace PingMonitor.Web.Services;
+
+internal sealed class PlaceholderStateEvaluationService : IStateEvaluationService
+{
+    public Task EvaluateAssignmentStateAsync(string assignmentId, CancellationToken cancellationToken)
+    {
+        return Task.CompletedTask;
+    }
+}
